Phase 4: Poteau isolé + Voile isolé (5-step load-descent flows)
Both flows share a stepper shell (StepperHeader/StepFooter — including
a 350ms debounce on Suivant/Terminer reproducing a real fix from the
prototype's iteration history) and a level-editor (LevelEditorCard):
storey height, slab type + thickness, EC1 usage category, and an
unlimited coatings list, each keyed by stable ids rather than list
position so removing an item from the middle can't leave a stale
TextEditingController showing another item's value.

Poteau (spec §4): système/règlement/béton → nœud with 4 tributary
spans (L1-L4, auto column-position badge) + beam sections when the
système includes poutres → niveaux → récapitulatif → résultats with
automatic section prédimensionnement.

Voile (spec §5): same shape plus a Vent EC1 block, a 2-span
(avant/arrière) tributary area instead of 4, and a cumulative wind
force per level with a "Vent dominant" badge once F_vent(ELU) exceeds
15% of N_ELU.

Domain layer changes to support this:
- New wind.dart: the prototype's simplified qp(z) = max(0.45,
  0.55+0.032z) pressure model, ported verbatim (zone/région/terrain
  are informational — the shipped prototype only varies pressure by
  elevation).
- load_descent.dart: split each level's G_dalle/G_rev/Q into their own
  per-level force (kN) versus the cumulative G_tot/Q_tot running
  totals — the previous shape conflated pressure (kN/m²) with a
  cumulative total, which would have rendered a wrong/redundant table
  matching neither unit nor the spec's 7 distinct columns. Added
  cumulateWallLoadDescent (elevation derived from storey heights,
  reusing the poteau accumulator's shape plus wind).

Extracted PredimResultPanel out of the prédim screen into a shared
widget so both flows' results steps reuse the exact same formulas and
result formatting as the standalone prédimensionnement screen.

Also fixes two latent `num`/`int` type risks from `.clamp()` (its
return type is `num` even on an `int`/`double` receiver) in
app_shell.dart and the new flow screens.

+++ lib/src/ui/shell/app_shell.dart
@@ -29,7 +29,7 @@ class _AppShellState extends State<AppShell> {
         const _Tab(label: "Mon compte", icon: Icons.person_outline, page: AccountScreen()),
     ];
 
-    final index = _index.clamp(0, tabs.length - 1);
+    final index = _index.clamp(0, tabs.length - 1).toInt();
 
     return Scaffold(
       body: IndexedStack(
